# backend/src/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Tapa Transport Platform API",
    "version": "1.2.0",
    "description": "REST API for Tapa's transport operations platform. It supports multi-role identity, agency operations, journeys, ticketing, parcels, encrypted wallets, bulk passenger manifests, waitlists, live trip events, and operational observability."
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local development"
    },
    {
      "url": "https://tapa.rw",
      "description": "Production"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT access token returned by the auth endpoints"
      }
    },
    "parameters": {
      "Page": {
        "name": "page",
        "in": "query",
        "schema": {
          "type": "integer",
          "minimum": 1,
          "default": 1
        }
      },
      "PageSize": {
        "name": "pageSize",
        "in": "query",
        "schema": {
          "type": "integer",
          "minimum": 1,
          "maximum": 100,
          "default": 20
        }
      },
      "IdempotencyKey": {
        "name": "Idempotency-Key",
        "in": "header",
        "required": true,
        "schema": {
          "type": "string",
          "minLength": 1
        },
        "description": "Required for wallet mutations and ticket mutations. Reusing a key with a different payload returns a conflict."
      }
    },
    "schemas": {
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          },
          "code": {
            "type": "string"
          },
          "status": {
            "type": "integer"
          },
          "requestId": {
            "type": "string"
          },
          "details": {
            "description": "Optional structured validation or domain details",
            "nullable": true
          }
        }
      },
...
      "Agency": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "ownerId": {
            "type": "string",
            "format": "uuid"
          },
          "verified": {
            "type": "boolean"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "Vehicle": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "plateNumber": {
            "type": "string"
          },
          "model": {
            "type": "string"
          },
          "capacity": {
            "type": "integer"
          },
          "agencyId": {
            "type": "string",
            "format": "uuid"
          },
          "driverId": {
            "type": [
              "string",
              "null"
            ]
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "Station": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "location": {
            "type": "string",
            "example": "-1.95,30.06"
          },
          "agencyId": {
            "type": "string",
            "format": "uuid"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "Journey": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "sourceStationId": {
            "type": "string",
            "format": "uuid"
          },
          "destinationStationId": {
            "type": "string",
            "format": "uuid"
          },
          "vehicleId": {
            "type": "string",
            "format": "uuid"
          },
          "departureTime": {
            "type": "string",
            "format": "date-time"
          },
          "price": {
            "type": "number"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "TicketPurchaseResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          },
          "ticket": {
            "$ref": "#/components/schemas/Ticket"
          },
          "remainingBalance": {
            "type": "number"
          }
        }
      },
...
      "ParcelWithJourney": {
        "allOf": [
          {
            "$ref": "#/components/schemas/Parcel"
          },
          {
            "type": "object",
            "properties": {
              "journey": {
                "type": "object",
                "properties": {
                  "id": {
                    "type": "string",
                    "format": "uuid"
                  },
                  "sourceStation": {
                    "type": "object",
                    "properties": {
                      "name": {
                        "type": "string"
                      }
                    }
                  },
                  "destinationStation": {
                    "type": "object",
                    "properties": {
                      "name": {
                        "type": "string"
                      }
                    }
                  }
                }
              }
            }
          }
        ]
      },
...
      "BulkBookingCreateResponse": {
        "type": "object",
        "required": [
          "bookingId",
          "masterQR",
          "passengers"
        ],
        "properties": {
          "bookingId": {
            "type": "string",
            "format": "uuid"
          },
          "masterQR": {
            "type": "string",
            "format": "uri",
            "description": "QR code returned as a PNG data URL produced by the QR helper",
            "example": "data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAA..."
          },
          "passengers": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BulkPassengerSummary"
            }
          }
        }
      },
...
      "CreateAgencyRequest": {
        "type": "object",
        "required": [
          "name",
          "ruraCode"
        ],
        "properties": {
          "name": {
            "type": "string"
          },
          "ruraCode": {
            "type": "string"
          }
        }
      },
...
      "CreateVehicleRequest": {
        "type": "object",
        "required": [
          "plateNumber",
          "model",
          "capacity",
          "agencyId"
        ],
        "properties": {
          "plateNumber": {
            "type": "string"
          },
          "model": {
            "type": "string"
          },
          "capacity": {
            "type": "integer",
            "minimum": 1
          },
          "agencyId": {
            "type": "string",
            "format": "uuid"
          }
        }
      },
      "CreateStationRequest": {
        "type": "object",
        "required": [
          "name",
          "location",
          "agencyId"
        ],
        "properties": {
          "name": {
            "type": "string"
          },
          "location": {
            "type": "string"
          },
          "agencyId": {
            "type": "string",
            "format": "uuid"
          }
        }
      },
      "CreateJourneyRequest": {
        "type": "object",
        "required": [
          "sourceStationId",
          "destinationStationId",
          "vehicleId",
          "departureTime",
          "price"
        ],
        "properties": {
          "sourceStationId": {
            "type": "string",
            "format": "uuid"
          },
          "destinationStationId": {
            "type": "string",
            "format": "uuid"
          },
          "vehicleId": {
            "type": "string",
            "format": "uuid"
          },
          "departureTime": {
            "type": "string",
            "format": "date-time"
          },
          "price": {
            "type": "number",
            "exclusiveMinimum": 0
          }
        }
      },
...
      "BuyTicketRequest": {
        "type": "object",
        "required": [
          "journeyId",
          "seatNumber",
          "walletPassword"
        ],
        "properties": {
          "journeyId": {
            "type": "string",
            "format": "uuid"
          },
          "seatNumber": {
            "type": "integer",
            "minimum": 1
          },
          "walletPassword": {
            "type": "string"
          }
        }
      },
...
      "CreateParcelRequest": {
        "type": "object",
        "required": [
          "journeyId",
          "receiverName",
          "receiverPhone"
        ],
        "properties": {
          "journeyId": {
            "type": "string",
            "format": "uuid"
          },
          "receiverName": {
            "type": "string"
          },
          "receiverPhone": {
            "type": "string"
          },
          "notes": {
            "type": "string"
          }
        }
      },
...
      "CreateBulkBookingRequest": {
        "type": "object",
        "required": [
          "organizationId",
          "journeyId",
          "destination",
          "departureTime",
          "passengers"
        ],
        "properties": {
          "organizationId": {
            "type": "string",
            "format": "uuid"
          },
          "journeyId": {
            "type": "string",
            "format": "uuid"
          },
          "destination": {
            "type": "string"
          },
          "departureTime": {
            "type": "string",
            "format": "date-time"
          },
          "passengers": {
            "type": "array",
            "minItems": 1,
            "items": {
              "type": "object",
              "required": [
                "name",
                "nationalId",
                "seatNumber"
              ],
              "properties": {
                "name": {
                  "type": "string"
                },
                "nationalId": {
                  "type": "string"
                },
                "seatNumber": {
                  "type": "integer",
                  "minimum": 1
                },
                "parentPhone": {
                  "type": "string"
                },
                "parentEmail": {
                  "type": "string",
                  "format": "email"
                }
              }
            }
          }
        }
      },
...
      "AddWaitlistRequest": {
        "type": "object",
        "required": [
          "journeyId",
          "boardingStopId"
        ],
        "properties": {
          "journeyId": {
            "type": "string",
            "format": "uuid"
          },
          "boardingStopId": {
            "type": "string",
            "format": "uuid"
          }
        }
      },
      "PaginatedUsers": {
        "allOf": [
          {
            "$ref": "#/components/schemas/PaginationMeta"
          },
          {
            "type": "object",
            "properties": {
              "items": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/UserSummary"
                }
              }
            }
          }
        ]
      },
      "PaginatedAgencies": {
        "allOf": [
          {
            "$ref": "#/components/schemas/PaginationMeta"
          },
          {
            "type": "object",
            "properties": {
              "items": {
                "type": "array",
                "items": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/Agency"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "owner": {
                          "type": "object",
                          "properties": {
                            "id": {
                              "type": "string",
                              "format": "uuid"
                            },
                            "email": {
                              "type": "string",
                              "format": "email"
                            }
                          }
                        },
                        "_count": {
                          "type": "object",
                          "properties": {
                            "vehicles": {
                              "type": "integer"
                            },
                            "stations": {
                              "type": "integer"
                            }
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        ]
      },
      "PaginatedVehicles": {
        "allOf": [
          {
            "$ref": "#/components/schemas/PaginationMeta"
          },
          {
            "type": "object",
            "properties": {
              "items": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/VehicleListItem"
                }
              }
            }
          }
        ]
      },
      "PaginatedStations": {
        "allOf": [
          {
            "$ref": "#/components/schemas/PaginationMeta"
          },
          {
            "type": "object",
            "properties": {
              "items": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/Station"
                }
              }
            }
          }
        ]
      },
      "PaginatedJourneys": {
        "allOf": [
          {
            "$ref": "#/components/schemas/PaginationMeta"
          },
          {
            "type": "object",
            "properties": {
              "items": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/JourneyListItem"
                }
              }
            }
          }
        ]
      },
...
    }
  },
  "paths": {
...
    "/mock/rura/verify": {
      "post": {
        "tags": [
          "Operations"
        ],
        "summary": "Mock RURA verification endpoint",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "ruraCode"
                ],
                "properties": {
                  "ruraCode": {
                    "type": "string",
                    "example": "RURA-12345"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Verification result",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "verified": {
                      "type": "boolean"
                    },
                    "code": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/users": {
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "List users",
        "description": "Accessible to OWNER and MANAGER roles.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/Page"
          },
          {
            "$ref": "#/components/parameters/PageSize"
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated users",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedUsers"
                }
              }
            }
          }
        }
      }
    },
    "/api/agencies": {
      "get": {
        "tags": [
          "Agencies"
        ],
        "summary": "List agencies",
        "parameters": [
          {
            "$ref": "#/components/parameters/Page"
          },
          {
            "$ref": "#/components/parameters/PageSize"
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated agencies",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedAgencies"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Agencies"
        ],
        "summary": "Create a verified agency",
        "description": "OWNER only. Calls the configured RURA verification service before persisting the agency.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAgencyRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Agency created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Agency"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/vehicles": {
      "get": {
        "tags": [
          "Vehicles"
        ],
        "summary": "List vehicles",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/Page"
          },
          {
            "$ref": "#/components/parameters/PageSize"
          },
          {
            "name": "agencyId",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated vehicles",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedVehicles"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Vehicles"
        ],
        "summary": "Create vehicle",
        "description": "OWNER only.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateVehicleRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Vehicle created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Vehicle"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/stations": {
      "get": {
        "tags": [
          "Stations"
        ],
        "summary": "List stations",
        "parameters": [
          {
            "$ref": "#/components/parameters/Page"
          },
          {
            "$ref": "#/components/parameters/PageSize"
          },
          {
            "name": "agencyId",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated stations",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedStations"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Stations"
        ],
        "summary": "Create station",
        "description": "OWNER only.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateStationRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Station created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Station"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/journeys": {
      "get": {
        "tags": [
          "Journeys"
        ],
        "summary": "List future journeys",
        "parameters": [
          {
            "$ref": "#/components/parameters/Page"
          },
          {
            "$ref": "#/components/parameters/PageSize"
          },
          {
            "name": "sourceId",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "destId",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated journeys",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedJourneys"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Journeys"
        ],
        "summary": "Create journey",
        "description": "MANAGER only.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateJourneyRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Journey created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Journey"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/tickets": {
      "post": {
        "tags": [
          "Tickets"
        ],
        "summary": "Buy ticket",
        "description": "CLIENT only. Requires a fully verified account, a wallet password, and an `Idempotency-Key` header.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/IdempotencyKey"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BuyTicketRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Ticket purchased",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TicketPurchaseResponse"
                }
              }
            }
          },
          "409": {
            "description": "Seat taken, temporarily reserved, or insufficient funds",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/parcels": {
      "post": {
        "tags": [
          "Parcels"
        ],
        "summary": "Create parcel",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateParcelRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Parcel created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ParcelWithJourney"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/bulk-bookings": {
      "post": {
        "tags": [
          "Bulk Bookings"
        ],
        "summary": "Create organization bulk booking manifest",
        "description": "Accessible to ORGANIZATION, MANAGER, and OWNER roles. Generates a master QR, passenger boarding hashes, and HMAC-backed manifest integrity data.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateBulkBookingRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Bulk booking created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BulkBookingCreateResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/waitlist": {
      "post": {
        "tags": [
          "Waitlist"
        ],
        "summary": "Add client to journey waitlist",
        "description": "CLIENT only. Requires a fully verified account. User joins a stop-specific FIFO queue backed by Redis sorted sets.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddWaitlistRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Waitlist entry created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WaitlistAddResponse"
                }
              }
            }
          }
        }
      }
    },
...
  }
}